🔨 refactor(ListViewFlatCard.tsx): remove unused props and simplify onPress function The 'i' prop is no longer used, so it has been removed. The onPress function has been simplified by passing only the 'id' prop to the 'flatShow' screen.
🔨 refactor(FlatInfoContainer.tsx): remove isLessor prop and commented out code
This commit removes the isLessor prop from the FlatInfoContainer component as it is no longer needed. It also removes commented out code that was previously used to conditionally render certain elements based on the value of isLessor.

🔨 refactor(HighlightButtons.tsx): extract pressHeart function to parent component
✨ feat(HighlightButtons.tsx): add onPressHeart and favorite props
The pressHeart function has been extracted to the parent component to improve separation of concerns and make the code more modular. The onPressHeart and favorite props have been added to allow for more flexibility in the usage of the component. The onPressHeart function is now passed down from the parent component and the favorite prop is used to determine whether the heart icon should be filled or not.

🔨 refactor(advertSlice.ts): add fromDate and toDate fields to AdvertState
✨ feat(advertSlice.ts): add fromDate and toDate fields to advert object
The AdvertState interface and initialState in advertSlice.ts have been updated to include fromDate and toDate fields. The advert object in the advertSlice reducer has also been updated to include these fields. This allows for more detailed information about the advert to be stored and displayed.

🔨 refactor(dateFormatConverter.ts): simplify date conversion logic
The dateFormatConverter function in dateFormatConverter.ts has been refactored to simplify the logic for converting dates. The if-else statement has been replaced with a ternary operator, which makes the code more concise and easier to read.

🔨 refactor(FlatShowScreen.tsx): change saveFlatsToFavorites to toggleFavorite
🚀 perf(FlatShowScreen.tsx): improve advert fetching and add date format converter helper
The saveFlatsToFavorites function has been replaced with toggleFavorite to better reflect the action being performed. The advert is now fetched using the advertMiddleware instead of the usersSlice. The dateFormatConverter helper has been added to format the date in the advert.

🔨 refactor(FlatShowScreen.tsx): refactor FlatShowScreen component
This commit refactors the FlatShowScreen component by changing the variable names to be more descriptive and consistent with the naming conventions. The save function is replaced with the toggleFavorite function, which toggles the favorite status of the advert. The flat object is replaced with the advert object, which contains the flat and price information. The fromDate and toDate properties are converted to a human-readable format using the dateFormatConverter function.