🔥 refactor(App.tsx): remove unused code and console.log statements This commit removes unused code and console.log statements from the App.tsx file. The useEffect hook now only dispatches the checkToken action when the authenticated state changes. The initializing state is now set to false only when the authenticated state changes. The return statement now renders the GuestStackNavigator when the user is not authenticated and the RootStack.Navigator when the user is authenticated.
🐛 fix(App.tsx): remove unnecessary conditional rendering
The conditional rendering for the landlord variable is removed as it is not necessary. The dashboard screen is always rendered, and the landlord screen is only rendered if the landlord variable is truthy.

🐛 fix(authenticationMiddleware.ts): add authorization header to signOut request
The signOut function now includes an authorization header with the token retrieved from EncryptedStorage. This ensures that the server can authenticate the request and perform the sign out operation.

🔥 refactor(authenticationSlice.ts): remove unused state properties
✨ feat(authenticationSlice.ts): simplify state management by using a single authenticated boolean property
The state properties id, email, and admin are no longer used in the authenticationSlice, so they have been removed. The state management has been simplified by using a single authenticated boolean property to indicate whether the user is authenticated or not. This reduces the complexity of the state and makes it easier to reason about.

🔥 refactor(flatMiddleware.ts): comment out unused fetchFlats function
The fetchFlats function is currently unused and has been commented out to improve code readability.