🔥 chore(App.tsx): remove commented out code 🚀 feat(App.tsx): add Google Signin configuration The commented out code has been removed to improve readability and maintainability of the codebase. Google Signin has been added to the application by configuring the GoogleSignin module with a webClientId and offlineAccess options.
🔒 chore(android): add Google Authentication support
The Google Authentication support has been added to the Android application. The necessary dependencies have been added to the build.gradle file, and the AndroidManifest.xml file has been updated to include the application package name. The signingConfigs for the debug and release build types have been updated to include the necessary credentials for Google Authentication.

🔧 chore(android): add Google Authentication classpath to build.gradle
🔒 chore(android): add MyApp Variables for Google Play to gradle.properties
The Google Authentication classpath has been added to the build.gradle file, which is required for Google Authentication to work. The MyApp Variables for Google Play have been added to the gradle.properties file, which are used to sign the app when uploading it to the Google Play Store.

🚀 feat(xcodeproj): add GoogleService-Info.plist to project resources
The GoogleService-Info.plist file has been added to the project resources. This file is required for Firebase integration in the iOS app.

🔍 chore(xcodeproj): add GoogleService-Info.plist to project
The GoogleService-Info.plist file was added to the project. This file is required for Firebase integration in the iOS app.

🔧 chore(ios): add GoogleService-Info.plist file and update Info.plist
🚀 feat(package.json): add @react-native-google-signin/google-signin dependency
The GoogleService-Info.plist file was added to the ios/lofftMobileApp directory. The Info.plist file was updated to use the correct reversed client ID. The @react-native-google-signin/google-signin dependency was added to the package.json file. This will allow the application to use Google Sign-In functionality.

🚀 feat(googleAuth.ts, SignInWith.tsx): add Google sign up functionality
The Google sign up functionality has been added to the application. The GoogleSignin module has been imported and configured with the webClientId and offlineAccess options. The GoogleSignUp function has been created to handle the sign in process. The function checks if the user has Google Play Services installed, then signs in the user and logs the result. If there is an error, it is logged to the console. The Google sign in button has been updated to call the GoogleSignUp function when pressed.