🔥 chore(ListViewFlatCard.tsx): remove redundant lines The code had two consecutive empty lines, which were removed to improve code readability.
🚀 feat(FlatInfoContainer.tsx): add advert object to component props
🔥 chore(FlatInfoContainer.tsx): remove unused variables and imports
The advert object is now passed as a prop to the FlatInfoContainer component. This allows the component to be more flexible and reusable as it can now display information for any advert object passed to it. Unused variables and imports have been removed to improve code readability and maintainability.

🐛 fix(FlatInfoContainer.tsx): fix references to advert object properties
✨ feat(FlatInfoContainer.tsx): add date formatting to fromDate and untilDate
The references to the advert object properties have been fixed. The properties were previously being accessed from the wrong object. The fromDate and untilDate properties are now formatted using the dateFormatConverter function. This improves the readability of the dates.

🎨 style(FlatInfoContainer.tsx): refactor FlatInfoContainer styles
💄 style(FlatInfoContainer.tsx): improve button style and add countdown timer
Refactored the styles of the FlatInfoContainer component to improve readability and maintainability. The button style was improved by adding a custom style and disabling the button if the user has already applied for the advert. A countdown timer was added to the component to show the time left before the application closes.

🐛 fix(advertMiddleware.ts): remove ContentType header from toggleFavorite request
✨ feat(advertMiddleware.ts): add applyForAdvert async thunk
The ContentType header is removed from the toggleFavorite request as it is not needed. The applyForAdvert async thunk is added to allow users to apply for an advert. It sends a POST request to the server with the advert id and the user's token in the Authorization header. If an error occurs, it is logged to the console.

✨ feat(advertSlice.ts): add applied field to AdvertState and update advert object
The applied field is added to the AdvertState to keep track of whether the user has applied for an advert or not. The advert object is updated to include the applied field.

🔥 chore(FlatShowScreen.tsx): remove unused imports and variables
This commit removes unused imports and variables from the FlatShowScreen component. The removed imports include Text, TouchableOpacity, FlatList, Image, IconButton, LofftIcon, fontStyles, CoreButton, and Chips. The removed variables include save.

🔨 refactor(FlatShowScreen.tsx): pass advert object to FlatInfoContainer
The FlatInfoContainer component now receives the advert object as a prop instead of individual properties. This simplifies the code and makes it more readable.