🔨 refactor(Chips.tsx): add tagSorter function and userData and type props 🔨 refactor(ListViewFlatCard.tsx): pass userProfile and advert.flat.characteristics to Chips component The tagSorter function has been added to the Chips component to sort the tags alphabetically. The userData and type props have been added to the Chips component to allow for more flexibility in displaying tags. The userProfile and advert.flat.characteristics props have been passed to the Chips component in the ListViewFlatCard component to display the tags for the specific flat.
🐛 fix(advertSlice.ts): fix typo in characteristics property name
✨ feat(advertSlice.ts): add characteristics and features properties to AdvertState
The characteristics property was misspelled as charachteristics, which has been fixed. Additionally, the AdvertState interface has been updated to include the characteristics and features properties, which are now also included in the initialState and the action payload. This allows for the storage and retrieval of these properties in the Redux store.

🔨 refactor(usersSlice.ts): add null to characteristics and filter properties
🆕 feat(tagSorter.ts): add tagSorter function
The characteristics and filter properties in the UserState interface now have a null value. This is to ensure that the properties are not undefined when accessed. The tagSorter function is a utility function that sorts tags into positive and negative arrays based on whether they are present in the userTags array.