🔨 refactor(AdvertMap.tsx): pass advert object to MapViewFlatCard component The MapViewFlatCard component now receives an advert object instead of individual properties. This simplifies the code and makes it more readable.
🎨 style(Chips.tsx): refactor ListChips component to accept more props
This commit refactors the ListChips component to accept more props such as emoji, positive, and features. The emoji prop is used to display an emoji next to the tag name, the positive prop is used to determine the color of the tag, and the features prop is used to determine the style of the tag. The component is now more flexible and can be used in more scenarios.

🎨 style(Chips.tsx): rename styles for better clarity
The styles for the component have been renamed to better reflect their purpose. The `propertyChip` style has been renamed to `chip` and the `personalityChip` style has been renamed to `characteristicTag`. Additionally, the font styles for the `propertyChip` and `personalityChip` have been renamed to `featureTagFont` and `characteristicTagFont` respectively.

🔀 refactor(ListViewFlatCard.tsx): extract tag sorting logic into helper function
This commit extracts the tag sorting logic into a helper function called `tagSorter`. The function takes two arrays of tags and returns two arrays of tags, one containing the tags that match between the two arrays and one containing the tags that do not match. The function is used to sort the user's characteristics and filter against the flat's characteristics and features. The `Chips` component is updated to use the sorted tags.

🔨 refactor(MapViewFlatCard.tsx): refactor MapViewFlatCard component
The MapViewFlatCard component has been refactored to improve readability and maintainability. The component now uses the tagSorter helper function to sort the positive and negative characteristics and features of the user and the advert. The advertR parameter has been renamed to advert and the id parameter has been removed. The favorite property has been renamed to advert.favorite and the match property has been renamed to advert.matchScore. The images property has been replaced with advert.flat.photos.

🎨 style(MapViewFlatCard.tsx): refactor Chips component props
The Chips component now receives two props: tags and features. The tags prop is used to display the tags of the advert, while the features prop is used to display the positive characteristics of the flat. This improves the readability of the code and makes it easier to understand the purpose of the Chips component.

🔥 refactor(tagSorter.ts): remove unused code and import statements
The import statement for `useAppSelector` is no longer used in the file, and the commented-out code for `userTags` is also not used. Therefore, they have been removed to improve code readability and maintainability.