🔥 chore(ListViewApplicantCard.tsx): remove unused Chips component 💄 style(ListViewApplicantCard.tsx): add temporary Text component for testing purposes The Chips component was commented out and not being used, so it was removed to improve code readability. A temporary Text component was added for testing purposes.
🔨 refactor(ListViewFlatCard.tsx): simplify tags destructuring and navigation parameter
This commit simplifies the destructuring of the tags from the tagSorter function, which now returns an object with positiveTags and negativeTags properties. The navigation parameter in the onPress function of the CoreButton component is also simplified to pass the entire advert object instead of just the id property.

🔨 refactor(MapViewFlatCard.tsx): simplify characteristics and features tags variables
The variables positiveCharacteristics and negativeCharacteristics have been refactored into a single variable characteristicsTags. Similarly, the variables positiveFeatures and negativeFeatures have been refactored into a single variable featuresTags. This simplifies the code and makes it easier to read.

🎨 style(FlatInfoContainer.tsx): add tags for flat characteristics and features
🔨 refactor(tagSorter.ts): change return value to an object
The FlatInfoContainer component now receives two new props, characteristicsTags and featuresTags, which are used to display tags for the flat's characteristics and features. The tagSorter helper function has been refactored to return an object instead of an array, which improves readability and makes it easier to access the positive and negative tags.

🔨 refactor(FlatShowScreen.tsx): remove unused state and add tagSorter helper
✨ feat(FlatShowScreen.tsx): add characteristicsTags and featuresTags props to FlatInfoContainer
The unused state 'matches' has been removed. The tagSorter helper has been added to sort the tags in the FlatInfoContainer. The characteristicsTags and featuresTags props have been added to the FlatInfoContainer to display the sorted tags.