📦 chore(Gemfile): add activesupport gem and specify version constraints 🔧 fix(project.pbxproj): add linker flags for d_classic library The Gemfile has been updated to include the activesupport gem with version constraints. This ensures that the application uses a version of activesupport that is greater than or equal to 6.1.7.3 and less than 7.1.0.
In the project.pbxproj file, linker flags have been added to include the d_classic library. This is achieved by adding the flag "-ld_classic" to the OTHER_LDFLAGS build setting.

🔧 chore(package.json): update react-native and metro-react-native-babel-preset versions
⬆️ Upgrade react-native version from 0.71.7 to 0.71.14 and metro-react-native-babel-preset version from 0.73.9 to 0.73.10
This commit updates the versions of react-native and metro-react-native-babel-preset in the package.json file. The new versions provide bug fixes, improvements, and compatibility updates.